binarysearchtree.cpp
This code will help you in creating the binary search tree and it will also traverse in all the 3 orders i,e., Inorder, Preorder and Postorder. While creating the BST we won't have the need to provide the direction for the node, as it is a BST so it will automatically put the lower node in the left side and the upper node in the right side.